+++ spring-boot-autoconfigure/src/main/java/org/axonframework/springboot/autoconfig/AxonAutoConfiguration.java
@@ -23,8 +23,8 @@ import org.axonframework.commandhandling.LoggingDuplicateCommandHandlerResolver;
 import org.axonframework.commandhandling.SimpleCommandBus;
 import org.axonframework.commandhandling.gateway.CommandGateway;
 import org.axonframework.commandhandling.gateway.DefaultCommandGateway;
+import org.axonframework.commandhandling.gateway.DefaultReactorCommandGateway;
 import org.axonframework.commandhandling.gateway.ReactorCommandGateway;
-import org.axonframework.commandhandling.gateway.ReactiveCommandGateway;
 import org.axonframework.common.transaction.TransactionManager;
 import org.axonframework.config.Configuration;
 import org.axonframework.config.EventProcessingConfigurer;
@@ -208,19 +208,19 @@ public class AxonAutoConfiguration implements BeanClassLoaderAware {
     @Bean
     @ConditionalOnMissingBean
     @ConditionalOnClass(name = "reactor.core.publisher.Mono")
-    public ReactiveCommandGateway reactiveCommandGateway(CommandBus commandBus) {
-        return ReactorCommandGateway.builder()
-                                    .commandBus(commandBus)
-                                    .build();
+    public ReactorCommandGateway reactiveCommandGateway(CommandBus commandBus) {
+        return DefaultReactorCommandGateway.builder()
+                                           .commandBus(commandBus)
+                                           .build();
     }
 
     @Bean
     @ConditionalOnMissingBean
     @ConditionalOnClass(name = "reactor.core.publisher.Mono")
-    public ReactiveQueryGateway reactiveQueryGateway(QueryBus queryBus) {
-        return ReactorQueryGateway.builder()
-                                  .queryBus(queryBus)
-                                  .build();
+    public ReactorQueryGateway reactiveQueryGateway(QueryBus queryBus) {
+        return DefaultReactorQueryGateway.builder()
+                                         .queryBus(queryBus)
+                                         .build();
     }
 
     @Bean
